+++ src/test/java/de/codescape/jira/plugins/scrumpoker/service/EstimationFieldServiceTest.java
@@ -5,7 +5,9 @@ import com.atlassian.jira.issue.IssueManager;
 import com.atlassian.jira.issue.MutableIssue;
 import com.atlassian.jira.issue.UpdateIssueRequest;
 import com.atlassian.jira.issue.fields.CustomField;
+import com.atlassian.jira.permission.ProjectPermissions;
 import com.atlassian.jira.security.JiraAuthenticationContext;
+import com.atlassian.jira.security.PermissionManager;
 import com.atlassian.jira.user.ApplicationUser;
 import de.codescape.jira.plugins.scrumpoker.model.GlobalSettings;
 import org.junit.Before;
@@ -39,6 +41,9 @@ public class EstimationFieldServiceTest {
     private CustomFieldManager customFieldManager;
 
     @Mock
+    private PermissionManager permissionManager;
+
+    @Mock
     private ScrumPokerSettingService scrumPokerSettingService;
 
     @InjectMocks
@@ -61,22 +66,63 @@ public class EstimationFieldServiceTest {
         when(jiraAuthenticationContext.getLoggedInUser()).thenReturn(applicationUser);
         when(issueManager.getIssueByCurrentKey(ISSUE_KEY)).thenReturn(issue);
         when(scrumPokerSettingService.load()).thenReturn(globalSettings);
-        when(globalSettings.getStoryPointField()).thenReturn(CUSTOM_FIELD_ID);
-        when(customFieldManager.getCustomFieldObject(CUSTOM_FIELD_ID)).thenReturn(customField);
     }
 
     @Test
     public void withErrorsDuringUpdateItShouldReturnFalse() {
-        when(issueManager.updateIssue(eq(applicationUser), eq(issue), any(UpdateIssueRequest.class)))
-            .thenThrow(new RuntimeException());
+        expectCustomFieldFound();
+        expectUpdatingIssueFails();
         assertThat(estimationFieldService.save(ISSUE_KEY, ESTIMATION), is(false));
     }
 
     @Test
     public void withSuccessfulUpdateItShouldReturnTrue() {
+        expectCustomFieldFound();
+        expectUpdatingIssueSuccessful();
+        assertThat(estimationFieldService.save(ISSUE_KEY, ESTIMATION), is(true));
+    }
+
+    @Test
+    public void withPermissionCheckEnabledAndNoPermission() {
+        expectPermissionCheckEnabled();
+        expectPermissionCheckFails();
+        assertThat(estimationFieldService.save(ISSUE_KEY, ESTIMATION), is(false));
+    }
+
+    @Test
+    public void withPermissionCheckEnabledAndPermissionExists() {
+        expectPermissionCheckEnabled();
+        expectPermissionCheckSuccessful();
+        expectCustomFieldFound();
+        expectUpdatingIssueSuccessful();
+        assertThat(estimationFieldService.save(ISSUE_KEY, ESTIMATION), is(true));
+    }
+
+    private void expectCustomFieldFound() {
+        when(globalSettings.getStoryPointField()).thenReturn(CUSTOM_FIELD_ID);
+        when(customFieldManager.getCustomFieldObject(CUSTOM_FIELD_ID)).thenReturn(customField);
+    }
+
+    private void expectPermissionCheckFails() {
+        when(permissionManager.hasPermission(ProjectPermissions.EDIT_ISSUES, issue, applicationUser)).thenReturn(false);
+    }
+
+    private void expectPermissionCheckSuccessful() {
+        when(permissionManager.hasPermission(ProjectPermissions.EDIT_ISSUES, issue, applicationUser)).thenReturn(true);
+    }
+
+    private void expectUpdatingIssueSuccessful() {
         when(issueManager.updateIssue(eq(applicationUser), eq(issue), any(UpdateIssueRequest.class)))
             .thenReturn(issue);
-        assertThat(estimationFieldService.save(ISSUE_KEY, ESTIMATION), is(true));
+    }
+
+    private void expectUpdatingIssueFails() {
+        when(issueManager.updateIssue(eq(applicationUser), eq(issue), any(UpdateIssueRequest.class)))
+            .thenThrow(new RuntimeException());
+    }
+
+    private void expectPermissionCheckEnabled() {
+        when(globalSettings.isCheckPermissionToSaveEstimate()).thenReturn(true);
     }
 
 }